FAC-907: Move AdviceItem enums to common module
