+++ flickit-advice-engine/src/main/java/org/flickit/assessment/advice/application/service/adviceitem/GetAdviceItemListService.java
@@ -1,7 +1,7 @@
 package org.flickit.assessment.advice.application.service.adviceitem;
 
 import lombok.RequiredArgsConstructor;
-import org.flickit.assessment.advice.application.domain.adviceitem.AdviceItem;
+import org.flickit.assessment.advice.application.domain.AdviceItem;
 import org.flickit.assessment.advice.application.port.in.adviceitem.GetAdviceItemListUseCase;
 import org.flickit.assessment.advice.application.port.out.adviceitem.LoadAdviceItemListPort;
 import org.flickit.assessment.advice.application.port.out.assessmentresult.LoadAssessmentResultPort;
