+++ flickit-advice-engine/src/test/java/org/flickit/assessment/advice/test/fixture/application/AdviceItemMother.java
@@ -1,9 +1,9 @@
 package org.flickit.assessment.advice.test.fixture.application;
 
-import org.flickit.assessment.advice.application.domain.adviceitem.AdviceItem;
-import org.flickit.assessment.advice.application.domain.adviceitem.CostLevel;
-import org.flickit.assessment.advice.application.domain.adviceitem.ImpactLevel;
-import org.flickit.assessment.advice.application.domain.adviceitem.PriorityLevel;
+import org.flickit.assessment.advice.application.domain.AdviceItem;
+import org.flickit.assessment.common.application.domain.adviceitem.CostLevel;
+import org.flickit.assessment.common.application.domain.adviceitem.ImpactLevel;
+import org.flickit.assessment.common.application.domain.adviceitem.PriorityLevel;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
