+++ flickit-assessment-common/src/main/java/org/flickit/assessment/common/application/domain/adviceitem/ImpactLevel.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.advice.application.domain.adviceitem;
+package org.flickit.assessment.common.application.domain.adviceitem;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Getter;
