+++ flickit-assessment-common/src/main/resources/i18n/advice/messages_en.properties
@@ -63,15 +63,3 @@ update-advice-item.assessmentResult.notFound=No assessmentResult found by this '
 delete-advice-item.id.notNull='adviceId' may not be null
 delete-advice-item.adviceItem.notFound=No adviceItem found by this 'adviceItemId'
 delete-advice-item.assessmentResult.notFound=No assessmentResult found by this 'assessmentId'
-
-CostLevel_LOW=Low
-CostLevel_MEDIUM=Medium
-CostLevel_HIGH=High
-
-PriorityLevel_LOW=Low
-PriorityLevel_MEDIUM=Medium
-PriorityLevel_HIGH=High
-
-ImpactLevel_LOW=Low
-ImpactLevel_MEDIUM=Medium
-ImpactLevel_HIGH=High
