+++ flickit-assessment-common/src/main/resources/i18n/advice/messages_fa.properties
@@ -63,15 +63,3 @@ update-advice-item.assessmentResult.notFound=هیچ نتیجه ارزیابی ب
 delete-advice-item.id.notNull='شناسه توصیه' نباید خالی باشد.
 delete-advice-item.adviceItem.notFound=هیچ توصیه‌ای با این 'شناسه توصیه' یافت نشد.
 delete-advice-item.assessmentResult.notFound=هیچ نتیجه ارزیابی با این 'شناسه ارزیابی' یافت نشد.
-
-CostLevel_LOW=کم
-CostLevel_MEDIUM=متوسط
-CostLevel_HIGH=زیاد
-
-PriorityLevel_LOW=کم
-PriorityLevel_MEDIUM=متوسط
-PriorityLevel_HIGH=زیاد
-
-ImpactLevel_LOW=کم
-ImpactLevel_MEDIUM=متوسط
-ImpactLevel_HIGH=زیاد
