+++ flickit-assessment-common/src/main/resources/i18n/messages_en.properties
@@ -39,3 +39,12 @@ invite-to-register.email.body.without.support.email=\
     <br>\
     <p>We await your presence.</p>\
     <p>Best, Team {1}</p>
+CostLevel_LOW=Low
+CostLevel_MEDIUM=Medium
+CostLevel_HIGH=High
+PriorityLevel_LOW=Low
+PriorityLevel_MEDIUM=Medium
+PriorityLevel_HIGH=High
+ImpactLevel_LOW=Low
+ImpactLevel_MEDIUM=Medium
+ImpactLevel_HIGH=High
