+++ flickit-assessment-common/src/main/resources/i18n/messages_fa.properties
@@ -39,3 +39,12 @@ invite-to-register.email.body.without.support.email=\
     <br>\
     <p>منتظر حضور شما هستیم.</p>\
     <p>با احترام، تیم {1}</p>
+CostLevel_LOW=کم
+CostLevel_MEDIUM=متوسط
+CostLevel_HIGH=زیاد
+PriorityLevel_LOW=کم
+PriorityLevel_MEDIUM=متوسط
+PriorityLevel_HIGH=زیاد
+ImpactLevel_LOW=کم
+ImpactLevel_MEDIUM=متوسط
+ImpactLevel_HIGH=زیاد
