+++ flickit-assessment-common/src/test/java/org/flickit/assessment/common/application/domain/adviceitem/PriorityLevelTest.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.advice.application.domain.adviceitem;
+package org.flickit.assessment.common.application.domain.adviceitem;
 
 import org.junit.jupiter.api.Test;
